Update BTL SCB

+++ BTL-SCB/src/services/BoardService.js
@@ -2,12 +2,37 @@ const Board = require("../models/Board");
 
 class BoardService {
 
+    checkBoardId = async (id) => { 
+        try {
+            const board = await Board.findOne({_id: id}) 
+
+            if(board == null) {
+                return false;
+            } else {
+
+                return board;
+            }
+          } catch (error) {
+            throw new Error('Không tồn tại BoardId này');
+          }
+        }
+
+
     create = async (data) =>{
         try {
             // Xử lý các nghiệp vụ liên quan
             // Gọi đến tầng model
             const board = new Board(data);
             await board.save();
+            return board;
+        } catch (error) {
+            throw error;
+        }
+    }
+
+    getNewBoard = async (id) => {
+        try {
+            const board = await Board.findById(id);
             return board;
         } catch (error) {
             throw error;
@@ -33,13 +58,30 @@ class BoardService {
         try {
             // Xử lý các nghiệp vụ liên quan
             // Gọi đến tầng model
-            const board = await Board.findById(id);
-            await board.deleteOne();
-            return true;
+            const checkBoardId = await Board.findOne({_id: id});
+            if (checkBoardId) {
+                const board = await Board.findById(id);
+                // console.log(board);
+                await board.deleteOne();
+                return true;
+            } else { 
+                return false;
+            }
+            
         } catch (error) {
-            throw error;
+            throw new Error('Không tồn tại BoardId này');
         }
     }
+
+    // delete = async (id) =>{
+    //     try {
+    //         const board = await Board.findById(id);
+    //         await board.deleteOne();
+    //         return true;
+    //     } catch (error) {
+    //         throw new Error('Không tồn tại BoardId này');
+    //     }
+    // }
 
 
     getAll = async () =>{
